// src/graphql/store/fragments/__generated__/ZoneFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: ZoneFragment
// ====================================================

export interface ZoneFragment {
  __typename: "Zone";
  id: string;
  city: string;
  zipCode: string;
}